+++ views/purchase/index.php
@@ -25,9 +25,11 @@ Helper::validateSession();
                         <td><?= htmlspecialchars($purchase->delivered_at()) ?? 'offen' ?></td>
                         <td><?= htmlspecialchars($purchase->status()) ?></td>
                         <td>
+                        <?php if (User::isAdmin()): ?>
                             <a href="./index.php?purchase/edit/<?= htmlspecialchars($purchase->id()) ?>" class="btn btn-warning btn-sm">Bearbeiten</a>
                             <a href="./index.php?purchase/delete/<?= htmlspecialchars($purchase->id()) ?>" class="btn btn-danger btn-sm">Löschen</a>
                             <a href="./index.php?purchase/update/<?= htmlspecialchars($purchase->id()) ?>" class="btn btn-success btn-sm">Geliefert</a>
+                        <?php endif; ?>
                         </td>
                     </tr>
                 <?php endforeach; ?>
